pyproject.toml
[project]
name = "docslm"
version = "0.1.0"
description = ""
authors = [
    {name = "Paolo Litta",email = "plitta@brembanarolle.com"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "django (>=6.0.1,<7.0.0)",
    "openpyxl (>=3.1.5,<4.0.0)",
    "pandas (>=2.3.3,<3.0.0)",
    "pyyaml (>=6.0.3,<7.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
